requirements.txt
requests
pandas
billboard.py
python-dotenv
streamlit
plotly